implemented simple store
