+++ arcana/core/cli/tests/test_cli_store.py
@@ -19,7 +19,7 @@ def test_store_cli(cli_runner, work_dir):
             add,
             [
                 store_name,
-                "arcana.core.utils.testing.data:MockDataStore",
+                "arcana.core.utils.testing.data:SimpleStore",
                 STORE_URI,
                 "--user",
                 STORE_USER,
@@ -33,7 +33,7 @@ def test_store_cli(cli_runner, work_dir):
         assert result.exit_code == 0, show_cli_trace(result)
         assert "file - arcana.dirtree.data.file_system:FileSystem" in result.output
         assert (
-            f"{store_name} - arcana.core.utils.testing.data:MockDataStore"
+            f"{store_name} - arcana.core.utils.testing.data:SimpleStore"
             in result.output
         )
         assert "    server: " + STORE_URI in result.output
@@ -49,7 +49,7 @@ def test_store_cli_remove(cli_runner, work_dir):
             add,
             [
                 new_store_name,
-                "arcana.core.utils.testing.data:MockDataStore",
+                "arcana.core.utils.testing.data:SimpleStore",
                 STORE_URI,
                 "--user",
                 STORE_USER,
@@ -78,7 +78,7 @@ def test_store_cli_rename(cli_runner, work_dir):
             add,
             [
                 old_store_name,
-                "arcana.core.utils.testing.data:MockDataStore",
+                "arcana.core.utils.testing.data:SimpleStore",
                 STORE_URI,
                 "--user",
                 STORE_USER,
@@ -88,15 +88,13 @@ def test_store_cli_rename(cli_runner, work_dir):
         )
         # Check store is saved
         result = cli_runner(ls, [])
-        assert "i123 - arcana.core.utils.testing.data:MockDataStore" in result.output
+        assert "i123 - arcana.core.utils.testing.data:SimpleStore" in result.output
 
         cli_runner(rename, [old_store_name, new_store_name])
         # Check store is renamed
         result = cli_runner(ls, [])
-        assert (
-            "i123 - arcana.core.utils.testing.data:MockDataStore" not in result.output
-        )
-        assert "y456 - arcana.core.utils.testing.data:MockDataStore" in result.output
+        assert "i123 - arcana.core.utils.testing.data:SimpleStore" not in result.output
+        assert "y456 - arcana.core.utils.testing.data:SimpleStore" in result.output
 
 
 def test_store_cli_encrypt_credentials(cli_runner, work_dir):
@@ -109,7 +107,7 @@ def test_store_cli_encrypt_credentials(cli_runner, work_dir):
             add,
             [
                 store_name,
-                "arcana.core.utils.testing.data:MockDataStore",
+                "arcana.core.utils.testing.data:SimpleStore",
                 STORE_URI,
                 "--user",
                 STORE_USER,
